# public/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: 'My API'
  version: '1.0'
paths:
...
  '/api/order/{id}':
    get:
      tags:
        - Order
      summary: 'Hiển thị tất cả các đơn hàng của người dùng có ID'
      description: 'Trả về tất cả các đơn hàng của người dùng có ID được cung cấp.'
      parameters:
        -
          name: id
          in: query
          description: 'ID của người dùng'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Danh sách đơn hàng của người dùng đó'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '404':
          description: 'Không tìm thấy đơn hàng nào cho người dùng đã cung cấp'
      security:
        -
          bearer: []
components:
  schemas:
...
    Order:
      title: Order
      required:
        - id
        - user_id
        - total
        - order_status
        - pay_type
        - order_id_ref
        - created_at
        - updated_at
      properties:
        id:
          description: 'ID của đơn hàng'
          type: integer
          format: int64
        user_id:
          description: 'ID của người dùng'
          type: integer
          format: int64
        total:
          description: 'Tổng giá trị của đơn hàng'
          type: number
          format: double
        order_status:
          description: 'Trạng thái của đơn hàng'
          type: string
        pay_type:
          description: 'Ngân hàng thanh toán'
          type: string
        order_id_ref:
          description: 'Mã đơn hàng'
          type: string
        created_at:
          description: 'Thời điểm tạo đơn hàng'
          type: string
          format: date-time
        updated_at:
          description: 'Thời điểm cập nhật đơn hàng'
          type: string
          format: date-time
      type: object
  securitySchemes:
    bearer:
      type: apiKey
      description: 'Enter token in format (Bearer <token>)'
      name: Authorization
      in: header